+++ front/src/types/likes/index.ts
@@ -1,0 +1,5 @@
+export type Like = {
+  id: number;
+  pictureId: number;
+  userId: number;
+}
